1.0.6. beta
* 1)

* Fixed Splash Cannon not showing bonus attack range on the client.
* Fixed Heart not showing bonus hp regen on the client.
* Fixed Enrage Crystal not having bonus hp regen amp, bonus lifesteal amp and damage reduction while stunned.
* Fixed Bloodstone not showing bonus hp regen and bonus mana regen on the client.
* Fixed Warp Beast Latch not being cancelled when Warp Beast is hexxed or banished.
* Simplified IsFirstItemInInventory
* Improved custom illusion code slightly.

* 2)

* 3)

* Removed animations library
* Updated Absolute Purge in basenpc library.
* Updated custom illusions library to be more stable.
* Updated timers library to be more informative about the errors when in tools mode.
* Updated player_resource library to be more stable.
* Updated SuperStrongDispel function.
* Updated FindUnitsInTrapezoid function to be more stable.
* Updated CustomCleaveAttack function to be more stable and errors give more info now.

* 4)

* Sohei Dash now also slows attack speed.
* Sohei is now spell-immune during Dash.
* Sohei now disjoints projectiles when casting Dash.
* Sohei Dash is now dispellable (with Nullifier and other dispels that pierce spell immunity).
* Sohei Dash can now be interrupted with roots and leashes that pierce spell immunity.
* Sohei Dash no longer has scepter or shard upgrade.
* Sohei Flurry of Blows reworked - increased radius, increased duration, invulnerable and rooted while it lasts.
* Sohei W is now Guard - Knocks back enemies around you, removes most negative effects and reflects all spell and attack projectiles that would hit you for the next 2 seconds. Can be used while stunned but consumes 25% of Sohei's max health.
* Sohei Guard has a shard upgrade - Guard applies Ki Strike to all enemies around Sohei.
* Sohei Momentum passive now applies only crit damage.
* Sohei's linked ability with Momentum is now Ki Strike - Sohei emits energy that damages, slows and pushes all enemy units in front of him. If an enemy unit is pushed into a tree, building, or cliff then that unit will be stunned. If pushed unit is a hero and it collides with another enemy hero, both heroes will be stunned.
* Sohei Ki Strike works normally while disarmed.
* Sohei talents readjusted.
* Enrage Crystal now provides 20% damage reduction while stunned.

* 5)

* Pudge Flesh Heap now has seperate stacks fro hero deaths and creep deaths.
* Pudge Flesh Heap now works for kills not made by Pudge - deaths within 250 range.
* Stealth Assassin Death Ward scepter number of bounces is now infinite but cannot bounce to already bounced units.
* Stealth Assassin Death Ward no longer benefits from auras and buffs.
* Pudge Rot slow talent improved from 14% to 16%.
* Stealth Assassin Death Ward damage increased from 70/125/180 to 90/150/210.
* Stealth Assassin Death Ward base attack speed changed from 100 to 70.
* Stealth Assassin Death Ward base attack time changed from 0.3 to 0.25
* Stealth Assassin Death Ward melee/ranged attacks to destroy changed from 4/5 to 3/4.
* Stealth Assassin Death Ward Level 25 Talent increased from +60 Death Ward Damage to +120.
* Hopefully improved spell cast detection for Silencer's spells.

* 6)

* Stealth Assassin Death Ward hp increased from 160 to 720. (it shouldn't anyone except those that use kill hp threshold like Axe)
* Mana Eater Absorb Mana no longer kills strong illusions.
* Mana Eater EMP no longer kills strong illusions.
* Mana Eater Mana Shell no longer provides mana while Mana Eater is dead.
* Mana Eater Mana Shell no longer provides mana when killing items on the ground or when killing runes.
* Mana Eater Mana Shell now has a mana gain indicator.
* Paladin Holy Strike : Trying to improve the ready particle.
* Silencer Curse of the Silent: Fixed multiple debuffs not stacking.

* 7)

* Added beta version of Bakedanuki - new hero overriding Dark Willow
* Added alpha version of Sona - new hero overriding Naga Siren
* Added beta version of Sandra - new hero overriding Phantom Assassin
* Dark Terminator Global Blink renamed to Teleport
* Added new ability Pew Pew to Dark Terminator.
* Dark Terminator Level 10 talent changed from +20 movement speed to +15% evasion.
* Dark Terminator Level 15 Talent +125 Attack Range changed to -1.5s Terminate cast time.
* Dark Terminator Level 15 Talent +25% Confuse illusion damage changed to +1 Confuse illusion.
* Dark Terminator Level 20 Talent +1 Confuse illusion changed to +50% Confuse illusion damage.
* Dark Terminator Level 20 Talent -1.6s Terminate cast time changed to +200 Attack Range.
* Dark Terminator Level 25 Talent +15% Terminator damage proc chance increased to +25%.
* Dark Terminator Terminate: Trying to fix Lotus Orb (reflect) interaction.
* Death Knight Death Pit now uses lifesteal and spell lifesteal instead of normal healing.

* 8)

* Blood Mage Phoenix egg hp increased from 160 to 720. (it shouldn't affect anyone except those that use kill hp threshold like Axe)
* Archmage shard spell immunity will now apply a basic dispel.
* Astral Trekker Giant Growth: Improved code and reduced lag.
* Brewmaster Drunken Haze now applies first tick of damage immediately when ignited.

* 9)

* Update .luacheckrc

* Update package-lock.json

* semistandard

* 7.32c + 7.32d

* Archmage min base damage increased from 36 to 38. (starting min damage increased from 50 to 52)
* Blademaster's Level 15 Talent Counter Attack proc chance increased from +10% to +12%.
* Blademaster Level 20 Talent +75 Omnislash cast range changed to +25% Counter Attack lifesteal.
* Dark Ranger base movement speed increased from 295 to 300.
* Silencer Curse of the Silent radius reduced from 425 to 350/375/400/425.
* Orb of Reflection active no longer interrupts channeling when used.
* Orb of Reflection no longer stacks with Blademail in any way.
* Eternal Shroud now shares cooldown with Hood of Defiance and Infused Robe.

* hmm

* Bakedanuki is playable

* Mana Flare and Stoneskin

* After testing

* Tooltips

+++ game/scripts/vscripts/heroes/mana_eater/absorb_mana.lua
@@ -3,10 +3,10 @@ function AbsorbMana(keys)
 	local caster = keys.caster
 	local target = keys.target
 	local ability = keys.ability
-	
+
 	-- Checking if target has spell block, if target has spell block, there is no need to execute the spell
 	if not target:TriggerSpellAbsorb(ability) then
-		if target:IsIllusion() then
+		if target:IsIllusion() and not target:IsStrongIllusionCustom() then
 			target:Kill(ability, caster) -- This gives the kill credit to the caster
 		else
 			local target_mana = target:GetMana()
